+++ lib/src/solid/login.dart
@@ -1,6 +1,6 @@
 /// A widget to obtain a Solid token to access the user's POD.
 ///
-// Time-stamp: <Friday 2024-01-05 09:02:57 +1100 Graham Williams>
+// Time-stamp: <Sunday 2024-01-07 08:33:23 +1100 Graham Williams>
 ///
 /// Copyright (C) 2024, Software Innovation Institute, ANU.
 ///
@@ -31,19 +31,12 @@ library;
 
 import 'package:flutter/material.dart';
 
+import 'package:package_info_plus/package_info_plus.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-import 'package:solid/src/login/solid_authenticate.dart';
+import 'package:solid/src/solid/authenticate.dart';
 import 'package:solid/src/widgets/popup_warning.dart';
 import 'package:solid/src/widgets/show_animation_dialog.dart';
-
-// The default package version string as the version of the app.
-
-// TODO 20231229 gjw GET THE ACTUAL VERSION FROM pubspec.yaml WHICH WILL BE THE
-// APP'S VERSION NOT THE SOLID PACKAGE'S VERSION USING
-// package_info_plus. https://github.com/anusii/solid/issues/18
-
-const appVersion = 'Version 0.0.0';
 
 // Screen size support funtions to identify narrow and very narrow screens. The
 // width dictates whether the Login panel is laid out on the right with the app
@@ -66,26 +59,18 @@ bool isVeryNarrowScreen(BuildContext context) =>
 /// user's POD is required when the app requires access to the user's POD for
 /// any of its functionality.
 
-class SolidLogin extends StatelessWidget {
+class SolidLogin extends StatefulWidget {
   const SolidLogin({
-    // Include the literals here so that they are exposed through the docs,
-    // except for version which is obtained from the importing app's
-    // pubspec.yaml by default.
+    // Include the literals here so that they are exposed through the docs.
 
     required this.child,
     this.image =
         const AssetImage('assets/images/default_image.jpg', package: 'solid'),
     this.logo =
         const AssetImage('assets/images/default_logo.png', package: 'solid'),
-    this.panelBG = const Color(0xFFF2F4FC),
     this.title = 'LOG IN TO YOUR POD',
     this.webID = 'https://pods.solidcommunity.au',
     this.link = 'https://solidproject.org',
-    this.getpodFG = Colors.purple,
-    this.getpodBG = Colors.orange,
-    this.loginFG = Colors.white,
-    this.loginBG = Colors.teal,
-    this.version = appVersion,
     super.key,
   });
 
@@ -101,11 +86,6 @@ class SolidLogin extends StatelessWidget {
 
   final AssetImage logo;
 
-  /// The Login panel's background colour. The default background colour is a
-  /// very light grey as a sublte background.
-
-  final Color panelBG;
-
   /// The login text indicating what we are loging in to.
 
   final String title;
@@ -119,30 +99,35 @@ class SolidLogin extends StatelessWidget {
 
   final String link;
 
-  /// The foreground colour of the GET POD button.
-
-  final Color getpodFG;
-
-  /// The background colour of the GET POD button.
-
-  final Color getpodBG;
-
-  /// The foreground colour of the LOGIN button.
-
-  final Color loginFG;
-
-  /// The background colour of the LOGIN button.
-
-  final Color loginBG;
-
-  /// The version of the app. The default is obtained from the app's
-  /// pubspec.yaml.
-
-  final String version;
-
-  /// The widget to hand over to once authentication is complete.
+  /// The child widget after logging in.
 
   final Widget child;
+  @override
+  State<SolidLogin> createState() => _SolidLoginState();
+}
+
+class _SolidLoginState extends State<SolidLogin> {
+  // This string will hold the application version number.  Initially, it's an
+  // empty string because the actual version number will be obtained
+  // asynchronously from the app's package information.
+
+  String appVersion = '';
+
+  @override
+  void initState() {
+    super.initState();
+    _initPackageInfo();
+  }
+
+  // Fetch the package information.
+
+  Future<void> _initPackageInfo() async {
+    final info = await PackageInfo.fromPlatform();
+
+    setState(() {
+      appVersion = info.version;
+    });
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -151,7 +136,7 @@ class SolidLogin extends StatelessWidget {
 
     final loginBoxDecor = BoxDecoration(
       image: DecorationImage(
-        image: image,
+        image: widget.image,
         fit: BoxFit.cover,
       ),
     );
@@ -159,44 +144,33 @@ class SolidLogin extends StatelessWidget {
     // Text controller for the URI of the solid server to which an authenticate
     // request is sent.
 
-    final webIdController = TextEditingController()..text = webID;
+    final webIdController = TextEditingController()..text = widget.webID;
 
-    // A GET A POD button that when pressed will launch a browser to
-    // the releveant link with instructions to get a POD.
+    // Define a common style for the text of the two buttons, GET POD and LOGIN.
 
-    const buttonLetterSpacing = 2.0;
-    const buttonFontSize = 15.0;
-    const buttonFontWeight = FontWeight.bold;
-    const buttonPadding = EdgeInsets.all(20);
-    final buttonBorderRadius = BorderRadius.circular(10);
+    const buttonTextStyle = TextStyle(
+      fontSize: 15.0,
+      letterSpacing: 2.0,
+      fontWeight: FontWeight.bold,
+    );
 
-    final getPodButton = TextButton(
-      style: TextButton.styleFrom(
-        padding: buttonPadding,
-        backgroundColor: getpodBG,
-        shape: RoundedRectangleBorder(
-          borderRadius: buttonBorderRadius,
-        ),
-      ),
+    // The GET A POD button that when pressed will launch a browser to the
+    // releveant link from where a user can register for a POD on the Solid
+    // server. The default location is relative to the [webID], and is currently
+    // a fixed path but needs to be obtained from the server meta data, as was
+    // done in solid_auth through [getIssuer].
 
+    final getPodButton = ElevatedButton(
       // TODO 20231229 gjw NEED TO USE AN APPROACH TO GET THE RIGHT SOLID SERVER
       // REGISTRATION URL WHICH HAS CHANGED OVER SERVERS. PERHAPS IT IS NEEDED
       // TO BE OBTAINED FROM THE SERVER META DATA? CHECK WITH ANUSHKA. MIGRATE
       // getIssuer() FROM solid-auth PERHAPS WITH lauchIssuerReg() IF THERE IS A
       // REQUIREMENT FOR THAT TOO?
 
-      onPressed: () =>
-          launchUrl(Uri.parse('$webID/.account/login/password/register/')),
+      onPressed: () => launchUrl(
+          Uri.parse('${widget.webID}/.account/login/password/register/')),
 
-      child: Text(
-        'GET A POD',
-        style: TextStyle(
-          color: getpodFG,
-          letterSpacing: buttonLetterSpacing,
-          fontSize: buttonFontSize,
-          fontWeight: buttonFontWeight,
-        ),
-      ),
+      child: const Text('GET A POD', style: buttonTextStyle),
     );
 
     // A LOGIN button that when pressed will proceed to attempt to connect to
@@ -204,62 +178,114 @@ class SolidLogin extends StatelessWidget {
     // themselves. On return from the authentication, if successful, the class
     // provided child widget is instantiated.
 
-    final loginButton = TextButton(
-      style: TextButton.styleFrom(
-        padding: buttonPadding,
-        backgroundColor: loginBG,
-        shape: RoundedRectangleBorder(
-          borderRadius: buttonBorderRadius,
-        ),
-      ),
-
-      // For now 20231230 simply go to the provided child widget on tap of the
-      // LOGIN button until the authentication is implemented. This will allow
-      // parallel implmentation of the app's GUI.
-
+    final loginButton = ElevatedButton(
+      // style: TextButton.styleFrom(
+      //   shape: RoundedRectangleBorder(
+      //     borderRadius: buttonBorderRadius,
+      //   ),
+      // ),
       onPressed: () async {
-        await Navigator.pushReplacement(
-          context,
-          MaterialPageRoute(
-            builder: (context) => child,
-          ),
-        );
+        // Authenticate against the Solid server.
+
+        // Method to show busy animation requiring BuildContext.
+        //
+        // This approach of creating a local method will avoid the `flutter
+        // analyze` issue `use_build_context_synchronously`, identifying the use
+        // of a BuildContext across asynchronous gaps, without referencing the
+        // BuildContext after the async gap.
+
+        void showBusyAnimation() {
+          showAnimationDialog(
+            context,
+            7,
+            'Logging in...',
+            false,
+          );
+        }
+
+        showBusyAnimation();
+
+        // Perform the actual authentication by contacting the server at
+        // [WebID].
+
+        final authResult = await solidAuthenticate(widget.webID, context);
+
+        // Method to navigate to the child widget, requiring BuildContext.
+
+        void navigateToApp() {
+          Navigator.pushReplacement(
+            context,
+            MaterialPageRoute(builder: (context) => widget.child),
+          );
+        }
+
+        // Method to show auth failed popup, requiring BuildContext.
+
+        void showAuthFailedPopup() {
+          popupWarning(context, 'Authentication has failed!');
+        }
+
+        // Check that the authentication succeeded, and if so navigate to the
+        // app itself. If it failed then notify the user and stay on the
+        // SolidLogin page.
+
+        if (authResult != null) {
+          navigateToApp();
+        } else {
+          showAuthFailedPopup();
+        }
       },
-      child: Text(
-        'LOGIN',
-        style: TextStyle(
-          color: loginFG,
-          letterSpacing: buttonLetterSpacing,
-          fontSize: buttonFontSize,
-          fontWeight: buttonFontWeight,
-          // TODO 20240104 gjw WHY THE CHOICE OF THIS SPECIFIC FONT? THIS WILL
-          // OVERRIDE ANY THEMES AND SO COULD CAUSE THE BUTTON TO LOOK RATHER
-          // DIFFERENT TO EVERYTHING ELSE WITHOUT A USER BEING ABLE TO FIX IT?
-          // fontFamily: 'Poppins',
-        ),
-      ),
+      child: const Text('LOGIN', style: buttonTextStyle),
     );
 
     // An Information link that is displayed within the Login panel.
 
-    Widget linkTo(String link) => GestureDetector(
-          onTap: () => launchUrl(Uri.parse(link)),
-          child: Container(
-            margin: const EdgeInsets.only(right: 20),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.end,
-              children: [
-                const Text('Visit '),
-                SelectableText(
+    Widget linkTo(String link) => Container(
+          margin: const EdgeInsets.only(right: 20),
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.end,
+            children: [
+              const Text('Visit '),
+
+              // Use a GestureDetector to capture a double tap to open the URL,
+              // and then within the SelectableText capture the single tap to
+              // display the URL. A longer tap will then select the text,
+              // ensuring we ignore it from the GestureDetector point of view,
+              // so it won;t be treated as a tap. I did try a Listener, which is
+              // a lower-level widget for handling pointer events, which allows
+              // the SelectableText, as its child, to remain selectable while
+              // also responding to taps to launch the URL, but it will always
+              // open the URL onPointerUp and had no simple onDoubleTap access.
+
+              // TODO 20240106 gjw Put the async anonymous function to launch
+              // the URL into a named function and call it twice in the below
+              // rather than repeating the code. DRY principle.
+
+              GestureDetector(
+                onLongPress: () => {},
+                onDoubleTap: () async {
+                  if (await canLaunchUrl(Uri.parse(link))) {
+                    await launchUrl(Uri.parse(link));
+                  } else {
+                    throw 'Could not launch $link';
+                  }
+                },
+                child: SelectableText(
                   link,
-                  textAlign: TextAlign.right,
-                  style: TextStyle(
-                      fontSize: screenWidth(context) > 400 ? 15 : 13,
-                      color: Colors.blue,
-                      decoration: TextDecoration.underline),
+                  onTap: () async {
+                    if (await canLaunchUrl(Uri.parse(link))) {
+                      await launchUrl(Uri.parse(link));
+                    } else {
+                      throw 'Could not launch $link';
+                    }
+                  },
+                  style: const TextStyle(
+                    color: Colors.blue,
+                    decoration: TextDecoration.underline,
+                  ),
                 ),
-              ],
-            ),
+              ),
+            ],
           ),
         );
 
@@ -277,7 +303,7 @@ class SolidLogin extends StatelessWidget {
       height: boxTextHeight,
       child: Center(
         child: SelectableText(
-          version,
+          'Version $appVersion',
           style: const TextStyle(
             color: versionTextColor,
           ),
@@ -293,7 +319,7 @@ class SolidLogin extends StatelessWidget {
       child: Column(
         children: [
           Image(
-            image: logo,
+            image: widget.logo,
             width: 200,
           ),
           const SizedBox(
@@ -303,7 +329,7 @@ class SolidLogin extends StatelessWidget {
           const SizedBox(
             height: 50.0,
           ),
-          Text(title,
+          Text(widget.title,
               textAlign: TextAlign.center,
               style: const TextStyle(
                 fontWeight: FontWeight.bold,
@@ -342,7 +368,7 @@ class SolidLogin extends StatelessWidget {
           ),
           Align(
             alignment: Alignment.centerRight,
-            child: linkTo(link),
+            child: linkTo(widget.link),
           ),
           // Expand to the bottom of the login panel.
           Expanded(
@@ -370,8 +396,7 @@ class SolidLogin extends StatelessWidget {
           horizontal: loginPanelInset * screenWidth(context)),
       child: SingleChildScrollView(
         child: Card(
-          elevation: 5,
-          color: panelBG,
+          elevation: 50,
           shape:
               RoundedRectangleBorder(borderRadius: BorderRadius.circular(15)),
           child: loginPanelDecor, //actualChildEventually,