+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.0 2024-01-05
+
+* Authentication implemented.
+* SolidLogin() initial version fully functional.
+
 ## 0.2.1 2024-01-01
 
 * Initial implementation of SolidLogin widget with parameters.